Move auth module helper fns
