+++ workers/loc.api/sync/authenticator/helpers/generate-sub-user-name.js
@@ -1,0 +1,14 @@
+'use strict'
+
+module.exports = (user, isSubAccount, isSubUser) => {
+  const { masterUserId, username: uName } = { ...user }
+  const subAccountNameEnding = isSubAccount
+    ? '-sub-account'
+    : ''
+  const subUserNameEnding = isSubUser
+    ? `-sub-user-${masterUserId}`
+    : ''
+  const username = `${uName}${subAccountNameEnding}${subUserNameEnding}`
+
+  return username
+}
