+++ workers/loc.api/sync/authenticator/helpers/index.js
@@ -1,0 +1,11 @@
+'use strict'
+
+const generateSubUserName = require('./generate-sub-user-name')
+const pickProps = require('./pick-props')
+const pickSessionProps = require('./pick-session-props')
+
+module.exports = {
+  generateSubUserName,
+  pickProps,
+  pickSessionProps
+}
